+++ tests/integ/modin/groupby/test_groupby_series.py
@@ -100,13 +100,14 @@ def test_groupby_agg_series_dict_func_negative():
     index = native_pd.Index(["a", "b", "b", "a", "c"])
     index.names = ["grp_col"]
     series = pd.Series([3.5, 1.2, 4.3, 2.0, 1.8], index=index)
+    native_series = native_pd.Series([3.5, 1.2, 4.3, 2.0, 1.8], index=index)
 
     eval_snowpark_pandas_result(
         series,
-        series.to_pandas(),
+        native_series,
         lambda se: se.groupby(by="grp_col").agg({"x": "min"}),
         expect_exception=True,
-        expect_exception_match="Value for func argument in dict format is not allowed for SeriesGroupBy.",
+        expect_exception_match="nested renamer is not supported",
         expect_exception_type=SpecificationError,
         assert_exception_equal=True,
     )